+++ packages/hedron-gradio/src/hedron_gradio/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 from hedron_gradio.client import GradioClientAdapter, GradioEndpoint
 from hedron_gradio.errors import GradioRemoteError
 from hedron_gradio.hf import HuggingFaceVendorNode, hf_space_node